+++ multiqc/multiqc.py
@@ -176,6 +176,13 @@ logger = config.logger
 @click.option("-q", "--quiet", is_flag=True, help="Only show log warnings")
 @click.option("--profile-runtime", is_flag=True, help="Add analysis of how long MultiQC takes to run to the report")
 @click.option("--no-ansi", is_flag=True, help="Disable coloured log output")
+@click.option(
+    "--custom-css-file",
+    "custom_css_files",
+    type=click.Path(exists=True, readable=True),
+    multiple=True,
+    help="Custom CSS file to add to the final report",
+)
 @click.version_option(config.version, prog_name="multiqc")
 def run_cli(
     analysis_dir,
@@ -215,6 +222,7 @@ def run_cli(
     quiet,
     profile_runtime,
     no_ansi,
+    custom_css_files,
     **kwargs,
 ):
     # Main MultiQC run command for use with the click command line, complete with all click function decorators.
@@ -274,6 +282,7 @@ def run_cli(
         quiet=quiet,
         profile_runtime=profile_runtime,
         no_ansi=no_ansi,
+        custom_css_files=custom_css_files,
         kwargs=kwargs,
     )
 
@@ -320,6 +329,7 @@ def run(
     quiet=False,
     profile_runtime=False,
     no_ansi=False,
+    custom_css_files=(),
     kwargs={},
 ):
     """MultiQC aggregates results from bioinformatics analyses across many samples into a single report.
@@ -438,6 +448,8 @@ def run(
         config.exclude_modules = exclude
     if profile_runtime:
         config.profile_runtime = True
+    if custom_css_files:
+        config.custom_css_files.extend(custom_css_files)
     config.kwargs = kwargs  # Plugin command line options
 
     # Clean up analysis_dir if a string (interactive environment only)